requirements.txt
asgiref==3.11.0
Django==6.0
django-tinymce==5.0.0
sqlparse==0.5.5